Update nodes.js
Created isEmpty() and removeFromFront() SLL methods. isEmpty() returns true if the SLL is empty and returns false if it is not.  removeFromFront() ensures the list is not empty, and if it is, returns null. Otherwise, the method will remove the current head, set the new head to be the node after the previous head, and return the previous head as a result.