Create twoNumSum.js
This function checks if an array contains two values that add up to a target sum. This is done in O(N) time and space complexity, and the function returns an array with the two numbers that add up to the target sum. If the array does not contain two values that add up to the target sum, a statement is returned.